Update interview_code_challenge
My logic was like this: First, get f,g,h,i expressed as a-e so that there will be fewer loops. Then since the result (FGHI) is a four-digit number, A and C must not be equal to 8 and 9(since it will reach 5-digit number), so I make the range of a and c to (1,7). Furthermore, b and e cannot be 1 and 5 since 1 leads to the repetition of number ,so does 5.(I filter those whose i equals to 5 or 0 in the last). The other filters are to filter those whose f>10 and those who have repetition problems .